Placing reusable function to file util
